Allow formatting empty voices in SOFT and FULL modes.
Previously, the formatter calculated the total ticks for a set of voices
based on the number of ticks used in the first voice. However, in SOFT
and FULL modes, the first voice is allowed to be empty, leading to a
total tick count of 0 breaking the rendering of other non-empty voices.

With this change, the formatter's total tick count is taken from the
getTotalTicks() method of the first voice - which will be non-zero even
if there are no notes in the voice, allowing other non-empty voices to
be laid out appropriately.